Change pakages, move code from main, separate by folders
