+++ Lesson_1/Marathon/Competitors/Competitor.java
@@ -1,4 +1,4 @@
-package Lesson_1.Marathon;
+package Marathon.Competitors;
 
 public interface Competitor {
     void run(int dist);
