+++ Lesson_1/Marathon/Main.java
@@ -1,17 +1,13 @@
-package Lesson_1.Marathon;
+package Marathon;
+import Marathon.Competitors.Team;
+import Marathon.Obstacles.Course;
 
 public class Main {
     public static void main(String[] args) {
-        Competitor[] competitors = {new Human("Боб"), new Cat("Барсик"), new Dog("Бобик")};
-        Obstacle[] course = {new Cross(80), new Water(2), new Wall(1), new Cross(120)};
-        for (Competitor c : competitors) {
-            for (Obstacle o : course) {
-                o.doIt(c);
-                if (!c.isOnDistance()) break;
-            }
-        }
-        for (Competitor c : competitors) {
-            c.info();
-        }
+        Team team = new Team("Prostokvashino");
+        Course c = new Course();
+        c.doIt(team); // Просим команду пройти полосу
+        team.info(); // Показываем инфо о команде
+        team.showResults(); // Показываем инфо об успехах
     }
 }
