+++ Lesson_1/Marathon/Obstacles/Obstacle.java
@@ -1,4 +1,5 @@
-package Lesson_1.Marathon;
+package Marathon.Obstacles;
+import Marathon.Competitors.Competitor;
 
 public abstract class Obstacle {
     public abstract void doIt(Competitor competitor);
